+++ projects/hipfft/clients/hipfft_params.h
@@ -1256,8 +1256,6 @@ private:
         {
             return HIPFFT_SUCCESS;
         }
-        throw unimplemented_exception("jit callbacks not implemented");
-#if 0
         hipfftResult_t       ret{HIPFFT_INVALID_PLAN};
         hipfftXtCallbackType cbtype = HIPFFT_CB_UNDEFINED;
         switch(itype)
@@ -1362,7 +1360,6 @@ private:
                                          ? nullptr
                                          : store_jit_cb_state->get_raw_data_ptrs().data());
         return ret;
-#endif
     }
 
     // call hipfftCreate + hipfftMake* functions, inserting calls to